Delete Inventory Repository

+++ IMS.UseCases/Inventories/ViewInventoryByIdUseCase.cs
@@ -18,7 +18,7 @@ namespace IMS.UseCases.Inventories
             this.inventoryRepository = inventoryRepository;
         }
 
-        public async Task<Inventory> ExecuteAsync(int inventoryId)
+        public async Task<Inventory> ExecuteAsync(long inventoryId)
         {
             return await this.inventoryRepository.GetInventoryByIdAsync(inventoryId);
         }
